// frontend/src/app/core/api/model/payrollRecordRequest.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PayrollRecordRequestDto { 
    amount: number;
    paymentDate: string;
    periodStart?: string;
    periodEnd?: string;
    paymentMethod?: PayrollRecordRequestDto.PaymentMethodEnum;
    notes?: string;
}
export namespace PayrollRecordRequestDto {
    export const PaymentMethodEnum = {
        Cash: 'CASH',
        Card: 'CARD',
        Wolt: 'WOLT',
        BankTransfer: 'BANK_TRANSFER',
        Other: 'OTHER'
    } as const;
    export type PaymentMethodEnum = typeof PaymentMethodEnum[keyof typeof PaymentMethodEnum];
}